+++ packages/dart/lib/quanttide_data.dart
@@ -2,4 +2,7 @@ library;
 
 export 'src/models/pipeline.dart';
 export 'src/models/dataset.dart';
+export 'src/repositories/pipeline_repository.dart';
+export 'src/repositories/dataset_repository.dart';
 export 'src/blocs/pipeline_bloc.dart';
+export 'src/blocs/dataset_bloc.dart';