+++ packages/dart/lib/src/repositories/dataset_repository.dart
@@ -1,0 +1,5 @@
+import 'package:quanttide_data/quanttide_data.dart';
+
+abstract class DatasetRepository {
+  Future<Dataset> fetch(String id);
+}