+++ packages/dart/lib/src/repositories/pipeline_repository.dart
@@ -1,0 +1,5 @@
+import 'package:quanttide_data/quanttide_data.dart';
+
+abstract class PipelineRepository {
+  Future<Pipeline> fetch(String id);
+}